refactor: move class URI to shared
